backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.12.3
pydantic-settings==2.1.0

# AWS
boto3==1.34.34
botocore==1.34.34
aioboto3==12.3.0

# ML & Deep Learning
torch==2.2.0
transformers==4.37.2
scikit-learn==1.4.0
numpy==1.26.3
scipy==1.12.0

# Vector Database
pinecone-client==3.0.2

# Text Processing
nltk==3.8.1
spacy==3.7.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# MLOps
mlflow==2.10.2
sagemaker==2.207.1

# API & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Utilities
python-json-logger==2.0.7
aiofiles==23.2.1
httpx==0.26.0
tenacity==8.2.3

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
moto==5.0.0
faker==22.6.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
bandit==1.7.6

